+++ packages/server/server-runtime/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @modern-js/server-runtime
 
+## 3.2.1
+
+### Patch Changes
+
+- @modern-js/server-core@3.2.1
+- @modern-js/runtime-utils@3.2.1
+- @modern-js/types@3.2.1
+
+## 3.2.0
+
+### Patch Changes
+
+- @modern-js/server-core@3.2.0
+- @modern-js/runtime-utils@3.2.0
+- @modern-js/types@3.2.0
+
+## 3.1.5
+
+### Patch Changes
+
+- @modern-js/server-core@3.1.5
+- @modern-js/runtime-utils@3.1.5
+- @modern-js/types@3.1.5
+
 ## 3.1.4
 
 ### Patch Changes